fix(nitro): share h3 event context type augmentation
