+++ packages/nitro-server/src/runtime/context.ts
@@ -1,0 +1,3 @@
+import '@nuxt/nitro-server/augments'
+
+export type { NuxtRequestContext } from '@nuxt/nitro-server/augments'
