+++ packages/nitro-server/src/runtime/middleware/no-ssr.ts
@@ -1,5 +1,7 @@
 import { defineEventHandler } from 'nitro/h3'
 
+import '../context'
+
 const handler: ReturnType<typeof defineEventHandler> = defineEventHandler((event) => {
   if (event.req.headers.get('x-nuxt-no-ssr')) {
     event.context.nuxt ||= {}
